+++ src/effects/chrisPremades.js
@@ -1,7 +1,9 @@
+/* eslint-disable no-await-in-loop */
 /* eslint-disable no-continue */
 /* eslint-disable require-atomic-updates */
 import DICTIONARY from "../dictionary.js";
 import CompendiumHelper from "../lib/CompendiumHelper.js";
+import logger from "../logger.js";
 import SETTINGS from "../settings.js";
 
 const DDB_FLAGS_TO_REMOVE = [
@@ -53,7 +55,8 @@ export async function applyChrisPremadeEffect(document, type, folderName = null)
   if (!game.modules.get("chris-premades")?.active) return document;
   const compendiumName = SETTINGS.CHRIS_PREMADES_COMPENDIUM.find((c) => c.type === type)?.name;
   if (!compendiumName) return document;
-  const chrisName = chrisPremades.ddb.getItemName(document.name);
+  const ddbName = document.flags.ddbimporter?.originalName ?? document.name;
+  const chrisName = CONFIG.chrisPremades?.renamedItems[ddbName] ?? ddbName;
   const folderId = type === "monsterfeatures"
     ? await getFolderId(folderName ?? chrisName, type, compendiumName)
     : undefined;
@@ -79,6 +82,7 @@ export async function applyChrisPremadeEffect(document, type, folderName = null)
 
   setProperty(document, "flags.ddbimporter.effectsApplied", true);
   setProperty(document, "flags.ddbimporter.chrisEffectsApplied", true);
+  logger.debug(`Updated ${document.name} with a Chris effect`);
 
   return document;
 }
@@ -100,9 +104,76 @@ export async function applyChrisPremadeEffects(documents, compendiumItem = false
       type = "inventory";
     }
 
-    // eslint-disable-next-line no-await-in-loop
     doc = await applyChrisPremadeEffect(doc, type);
   }
 
   return documents;
+}
+
+export async function addAndReplaceRedundantChrisDocuments(actor, folderName = null) {
+  if (!game.modules.get("chris-premades")?.active) return;
+  logger.debug("Beginning additions and removals of extra effects.");
+  const documents = actor.getEmbeddedCollection("Item").toObject();
+  const toAdd = [];
+  const toDelete = [];
+
+  for (let doc of documents) {
+    let type = doc.type;
+
+    if (["class", "subclass", "background"].includes(type)) continue;
+    const compendiumName = SETTINGS.CHRIS_PREMADES_COMPENDIUM.find((c) => c.type === type)?.name;
+    if (!compendiumName) continue;
+    if (DICTIONARY.types.inventory.includes(doc.type)) {
+      type = "inventory";
+    }
+
+    const ddbName = doc.flags.ddbimporter?.originalName ?? doc.name;
+    const chrisName = CONFIG.chrisPremades?.renamedItems[ddbName] ?? ddbName;
+    const newItemNames = getProperty(CONFIG, `chrisPremades.additionalItems.${chrisName}`);
+
+    if (newItemNames) {
+      logger.debug(`Adding new items for ${chrisName}`);
+      const folderId = type === "monsterfeatures"
+        ? await getFolderId(folderName ?? chrisName, type, compendiumName)
+        : undefined;
+
+      for (const newItemName of newItemNames) {
+        logger.debug(`Adding new item ${newItemName}`);
+        const chrisDoc = await chrisPremades.helpers.getItemFromCompendium(compendiumName, newItemName, true, folderId);
+        if (!documents.find((d) => d.name === chrisDoc.name)) toAdd.push(chrisDoc);
+      }
+    }
+
+    const itemsToRemoveNames = getProperty(CONFIG, `chrisPremades.removedItems.${chrisName}`);
+    if (itemsToRemoveNames) {
+      logger.debug(`Removing items for ${chrisName}`);
+      for (const removeItemName of itemsToRemoveNames) {
+        logger.debug(`Removing item ${removeItemName}`);
+        const deleteDoc = documents.find((d) => d.name === removeItemName);
+        if (deleteDoc) toDelete.push(deleteDoc._id);
+      }
+    }
+  }
+
+  logger.debug("Final Chris's Premades list", {
+    toDelete,
+    toAdd,
+  });
+  await actor.deleteEmbeddedDocuments("Item", toDelete);
+  await actor.createEmbeddedDocuments("Item", toAdd);
+
+}
+
+export async function addChrisEffectsToActorDocuments(actor) {
+  if (!game.modules.get("chris-premades")?.active) {
+    ui.notifications.error("Chris Premades module not installed");
+    return;
+  }
+
+  logger.info("Starting to update actor documents with Chris Premades effects");
+  let documents = actor.getEmbeddedCollection("Item").toObject();
+  const data = await applyChrisPremadeEffects(documents, false, true);
+  await actor.updateEmbeddedDocuments("Item", data);
+  await addAndReplaceRedundantChrisDocuments(actor);
+  logger.info("Effect replacement complete");
 }