+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ```javascript
 actor = game.actors.getName("Zinroe");
-await game.modules.get("ddb-importer")?.api.effects.addChrisEffectsToActorDocuments(actor);
+await game.modules.get("ddb-importer")?.api.chris.adjustActor(actor);
 ```
 
 If you want to apply to all characters in your world:
@@ -19,7 +19,7 @@ If you want to apply to all characters in your world:
 ```javascript
 for (const [key, actor] of game.actors.entries()) {
   console.log('Updating: ' + actor.name);
-  await game.modules.get("ddb-importer").api.effects.addChrisEffectsToActorDocuments(actor);
+  await game.modules.get("ddb-importer").api.chris.adjustActor(actor);
 }
 ```